Cleanup old tests
