+++ aicoach/prompt.py
@@ -23,7 +23,6 @@ class LoggingEnvironment(Environment):
 class Jinja2Loader:
     env: Environment
     scanner: Template
-    tags: Template
     new_replay: Template
     summary: Template
     initial_instructions: Template
@@ -35,7 +34,6 @@ class Jinja2Loader:
             loader=FileSystemLoader(searchpath="./aicoach/prompts/")
         )
         self.scanner = self.env.get_template("scanner.jinja2")
-        self.tags = self.env.get_template("tags.jinja2")
         self.new_replay = self.env.get_template("new_replay.jinja2")
         self.summary = self.env.get_template("summary.jinja2")
         self.initial_instructions = self.env.get_template("initial_instructions.jinja2")
